language and pinnedapp improvements

+++ src/ts/process/instance.ts
@@ -59,6 +59,7 @@ export class Process {
   }
 
   async killSelf() {
+    if (this._disposed) return;
     this.Log(`Killing self (PID ${this.pid})`);
     await this.handler.kill(this.pid);
   }
